Fix quality and performance issues with GDI+ usage
The Bitmap was being created twice (once with Image.FromFile, and once with
the Bitmap constructor), resulting in two copies being created in memory.

Use the overload of Image.FromFile that specifies color management should
be performed to get correct interpretation of the embedded ICC profiles
in the test images.

Use the correct ImageAttributes and PixelOffsetMode values to eliminate
edge artifacts (light lines near edge).  See
http://photosauce.net/blog/post/image-scaling-with-gdi-part-3-drawimage-and-the-settings-that-affect-it